+++ src/main/java/dev/tobee/telegram/request/body/SendDiceBody.java
@@ -1,7 +1,7 @@
 package dev.tobee.telegram.request.body;
 
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.tobee.telegram.model.media.DiceEmoji;
@@ -12,7 +12,7 @@ public record SendDiceBody(
         @JsonProperty("emoji") Optional<DiceEmoji> emoji,
         @JsonProperty("disable_notification") Optional<Boolean> disableNotification,
         @JsonProperty("protect_content") Optional<Boolean> protectContent,
-        @JsonProperty("reply_to_message_id") OptionalInt replyToMessageId,
+        @JsonProperty("reply_to_message_id") OptionalLong replyToMessageId,
         @JsonProperty("allow_sending_without_reply") Optional<Boolean> allowSendingWithoutReply,
         @JsonProperty("reply_markup") Optional<ReplyMarkup> replyMarkup
 ) {
