+++ src/main/java/dev/tobee/telegram/request/body/SendLocationBody.java
@@ -2,6 +2,7 @@ package dev.tobee.telegram.request.body;
 
 import java.util.Optional;
 import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import dev.tobee.telegram.model.message.ReplyMarkup;
@@ -16,7 +17,7 @@ public record SendLocationBody(
         @JsonProperty("proximity_alert_radius") OptionalInt proximityAlertRadius,
         @JsonProperty("disable_notification") Optional<Boolean> disableNotification,
         @JsonProperty("protect_content") Optional<Boolean> protectContent,
-        @JsonProperty("reply_to_message_id") OptionalInt replyToMessageId,
+        @JsonProperty("reply_to_message_id") OptionalLong replyToMessageId,
         @JsonProperty("allow_sending_without_reply") Optional<Boolean> allowSendingWithoutReply,
         @JsonProperty("reply_markup") Optional<ReplyMarkup> replyMarkup
 ) {
