+++ src/test/java/dev/tobee/telegram/request/message/SendMessageTest.java
@@ -2,7 +2,7 @@ package dev.tobee.telegram.request.message;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import dev.tobee.telegram.model.message.Message;
@@ -17,7 +17,7 @@ class SendMessageTest {
     void checkValidRequest() {
         SendMessage sendMessage = new SendMessage(
                 new SendMessageBody(159L, "eval", Optional.empty(), List.of(), Optional.empty(), Optional.empty(),
-                        Optional.empty(), OptionalInt.empty(), Optional.empty(), Optional.empty())
+                        Optional.empty(), OptionalLong.empty(), Optional.empty(), Optional.empty())
         );
 
         Assertions.assertEquals(sendMessage.getResponseType().getType().getTypeName(),
