+++ src/test/java/dev/tobee/telegram/request/sendobject/SendDiceTest.java
@@ -1,7 +1,7 @@
 package dev.tobee.telegram.request.sendobject;
 
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import dev.tobee.telegram.model.media.DiceEmoji;
@@ -17,7 +17,7 @@ class SendDiceTest {
     void checkValidRequest() {
         SendDice sendDice = new SendDice(
                 new SendDiceBody(159L, Optional.of(DiceEmoji.DIRECT_HIT), Optional.empty(), Optional.empty(),
-                        OptionalInt.empty(),
+                        OptionalLong.empty(),
                         Optional.empty(), Optional.empty())
         );
