+++ src/test/java/dev/tobee/telegram/request/sticker/SendStickerTest.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import dev.tobee.telegram.model.media.InputFile;
@@ -29,7 +29,7 @@ class SendStickerTest {
         );
 
         SendSticker sendSticker = new SendSticker(new SendStickerBody(159L, inputFile, Optional.empty(),
-                Optional.empty(), OptionalInt.empty(), Optional.empty(), Optional.empty()));
+                Optional.empty(), OptionalLong.empty(), Optional.empty(), Optional.empty()));
 
         Assertions.assertEquals(sendSticker.getResponseType().getType().getTypeName(),
                 (new TypeReference<ResponseWrapper<Message>>() {
